requirements.txt
numpy==1.16.1
plotly==4.13.0
dash==1.17.0
dash-core-components==1.13.0
dash-cytoscape==0.2.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
dash-daq==0.5.0
networkx==2.5
pymongo==3.9.0
gunicorn==20.0.4
cached-property==1.4.2
Werkzeug==1.0.1
python-dateutil==2.7.3